Update index.html popup close X
Automatically change the text colour of the close button (white X) when it is over a white background.

The popup has an X in the upper right corner to close/dismiss the popup. On a desktop browser, this is white text over a black background. In a mobile browser, the spectrogram takes up the majority of the screen, and the white X is not visible when it is above the white background of the spectrogram.